--- src/uri/utils.ts
@@ -1,19 +1,0 @@
-import url from "node:url";
-
-/**
- * Given a URL, return the host in a format that is safe to write.
- */
-export function toSafeHost(rawUrl: string): string {
-	const u = new URL(rawUrl);
-	// If the host is invalid, an empty string is returned.  Although, `new URL`
-	// should already have thrown in that case.
-	return url.domainToASCII(u.hostname) || u.hostname;
-}
-
-export function removeTrailingSlashes(value: string): string {
-	let end = value.length;
-	while (end > 0 && value[end - 1] === "/") {
-		end--;
-	}
-	return end === value.length ? value : value.slice(0, end);
-}
